requirements.txt
absl-py==1.4.0
albumentations==1.3.1
apache-beam==2.48.0
asttokens==2.2.1
astunparse==1.6.3
backcall==0.2.0
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
crcmod==1.7
customtkinter==5.2.0
cycler==0.11.0
darkdetect==0.8.0
debugpy==1.6.7
decorator==5.1.1
dill==0.3.1.1
dnspython==2.3.0
docopt==0.6.2
executing==1.2.0
fastavro==1.7.4
fasteners==0.18
ffmpeg-python==0.2.0
filelock==3.12.1
flatbuffers==23.3.3
fonttools==4.39.3
future==0.18.3
gast==0.4.0
google-api-core==2.11.1
google-api-python-client==1.12.11
google-apitools==0.5.31
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.11.1
google-cloud-bigquery-storage==2.20.0
google-cloud-bigtable==2.17.0
google-cloud-core==2.3.2
google-cloud-datastore==2.15.2
google-cloud-dlp==3.12.1
google-cloud-language==2.10.0
google-cloud-pubsub==2.17.1
google-cloud-pubsublite==1.8.2
google-cloud-recommendations-ai==0.10.3
google-cloud-spanner==3.36.0
google-cloud-videointelligence==2.11.2
google-cloud-vision==3.4.2
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
graphviz==0.20.1
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.2
h5py==3.8.0
hdfs==2.7.0
httplib2==0.22.0
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.6.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.13.2
ipywidgets==8.0.6
jax==0.4.8
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-widgets==3.0.7
keras==2.10.0
Keras-Preprocessing==1.1.2
keras-unet-collection==0.1.13
kiwisolver==1.4.4
lazy_loader==0.2
libclang==16.0.0
llvmlite==0.40.1rc1
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
ml-dtypes==0.1.0
mpmath==1.3.0
MyQR==2.3.1
natsort==8.3.1
nest-asyncio==1.5.6
netaddr==0.8.0
netifaces==0.11.0
networkx==3.1
numba==0.57.0
numpy==1.22.4
oauth2client==4.1.3
oauthlib==3.2.2
objsize==0.6.1
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
orjson==3.9.1
overrides==6.5.0
packaging==23.1
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
proto-plus==1.22.2
protobuf==3.19.6
psutil==5.9.5
pure-eval==0.2.2
pyarrow==6.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydot==1.4.2
Pygments==2.15.1
pymongo==4.3.3
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
pywin32==306
PyYAML==6.0
pyzmq==25.0.2
qudida==0.0.4
regex==2023.6.3
requests==2.30.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
shapely==2.0.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.12
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-cpu==2.10.0
tensorflow-directml-plugin==0.4.0.dev230202
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-metadata==1.13.1
tensorflow-serving-api==2.12.1
tensorflow_intel==2.10.0
termcolor==2.3.0
tfx-bsl==1.13.0
threadpoolctl==3.1.0
tifffile==2023.4.12
tk==0.1.0
torch==2.0.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==2.0.2
wcwidth==0.2.6
Werkzeug==2.3.3
widgetsnbextension==4.0.7
wrapt==1.14.1
zipp==3.15.0
zstandard==0.21.0
